+++ R/brick_to_dt.R
@@ -6,23 +6,27 @@
 #' @importFrom methods as
 #' @importFrom raster as.data.frame
 #' @importFrom R.utils getAbsolutePath
-#' @param data_file a character string with the path to the data file.
+#' @param x a character string with the path to the data file.
 #' @return data.table
 #' @export
 #' @examples
 #' \dontrun{
-#' x <- nc_to_csv("gpcp_tp_mm_global_197901_202205_025_monthly.nc")
-#' y <- nc_to_csv("dummie.nc")
+#' x <- brick_to_dt("gpcp_tp_mm_global_197901_202205_025_monthly.nc")
+#' y <- brick_to_dt("dummie.nc")
 #' }
 
-nc_to_csv <- function(data_file){
-  nc_in <- getAbsolutePath(data_file)
-  checker <- name_check(data_file)
+brick_to_dt <- function(x){
+  nc_in <- getAbsolutePath(x)
+  checker <- name_check(x)
   if (checker$length != 8) {
     warning("This is not pRecipe data")
   }
   dummie_cols <- aux_ts(checker$name[1])
-  dummie_brick <- brick(nc_in)
+  if (is.character(x)){
+    dummie_brick <- brick(nc_in)
+  } else {
+    dummie_brick <- x
+  }
   dummie_table <- as.data.frame(dummie_brick, xy = TRUE, long = TRUE,
                                 na.rm = TRUE)
   dummie_table <- as.data.table(dummie_brick)