+++ R/plot_map.R
@@ -7,41 +7,43 @@
 #' @importFrom methods as is
 #' @importFrom raster as.data.frame
 #' @importFrom stats quantile
-#' @param dummie a .nc file with precipitation
+#' @param x a .nc file with precipitation
 #' @return ggplot object
 #' @export
 
-plot_map <- function(dummie){
-  if (!is(dummie, "RasterLayer")){
-    stop("Error: please plot one Raster layer at the time")
+plot_map <- function(x){
+  if (!is(x, 'data.table')){
+    if (!is(x, "RasterLayer")){
+      stop("Error: please plot one Raster layer at the time")
+    }
+    x <- as.data.frame(x, xy = TRUE, long = TRUE, na.rm = TRUE)
+    x <- as.data.table(x)
   }
-  dummie <- as.data.frame(dummie, xy = TRUE, long = TRUE, na.rm = TRUE)
-  dummie <- as.data.table(dummie)
-  if (round((max(dummie$x) - min(dummie$x))) == 360){
+  if (round((max(x$x) - min(x$x))) == 360){
     map_expand <- FALSE
   } else {
     map_expand <- TRUE
   }
-  map_max <- quantile(dummie$value,0.9995)
-  map_min <- round(min(dummie$value)) -1 
+  map_max <- quantile(x$value,0.9995)
+  map_min <- round(min(x$value)) -1 
   if (map_min < 0){
     map_min <- 0
   }
-  p00 <- ggplot(dummie, aes(x = x, y = y)) +
+  p00 <- ggplot(x, aes(x = x, y = y)) +
     geom_raster(aes(fill = value)) +
     borders(colour = "black") +
     theme_bw() +
-    coord_cartesian(xlim = c(min(dummie$x), max(dummie$x)), 
-                    ylim = c(min(dummie$y), max(dummie$y)),
+    coord_cartesian(xlim = c(min(x$x), max(x$x)), 
+                    ylim = c(min(x$y), max(x$y)),
                     expand = map_expand) +
-    labs(x = NULL, y = NULL, fill = "[mm]", title = format(as.Date(dummie$Z[1]), "%B %Y")) +
+    labs(x = NULL, y = NULL, fill = "[mm]", title = format(as.Date(x$Z[1]), "%B %Y")) +
     scale_fill_distiller(palette = "YlGnBu",
                          direction = 1,
                          na.value = "#081D58",
                          limits = c(map_min, map_max),
                          guide = guide_colorbar(frame.colour = "black",
                                                 ticks.colour = "black")) +
-    theme(panel.border = element_rect(colour = "black", size = 2),
+    theme(panel.border = element_rect(colour = "black", linewidth = 2),
           panel.grid = element_blank(), plot.tag = element_text(size = 24), 
           plot.title = element_text(size = 28),
           axis.text = element_text(size = 20),